octofit-tracker/backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
django-cors-headers==3.14.0
pymongo==4.3.0
python-dotenv==1.0.0
djangorestframework-simplejwt==5.2.2